requirements.txt
Django==6.0.1
djangorestframework>=3.14
gunicorn>=21.2
python-dotenv>=1.0
